backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
bcrypt
httpx
google-genai
playwright
Pillow
cloudinary
python-multipart
email-validator
